chore: move trace column def to shared server to use env imports
