+++ packages/shared/src/server/services/traces-ui-table-service.ts
@@ -1,5 +1,5 @@
 import { OrderByState } from "../../interfaces/orderBy";
-import { tracesTableUiColumnDefinitions } from "../../tableDefinitions";
+import { tracesTableUiColumnDefinitions } from "../tableMappings";
 import { FilterState } from "../../types";
 import {
   StringFilter,
