+++ packages/shared/src/server/tableMappings/index.ts
@@ -1,0 +1,5 @@
+export * from "./mapObservationsTable";
+export * from "./mapTracesTable";
+export * from "../../tableDefinitions/mapDashboards";
+export * from "./mapScoresTable";
+export * from "./mapDatasetRunItemsTable";
