+++ packages/shared/src/server/tableMappings/mapSessionTable.ts
@@ -1,4 +1,4 @@
-import { UiColumnMappings } from ".";
+import { UiColumnMappings } from "../../tableDefinitions";
 
 export const sessionCols: UiColumnMappings = [
   // we do not access the traces scores in ClickHouse. We default back to the trace timestamps.
