+++ packages/shared/src/tableDefinitions/index.ts
@@ -1,8 +1,4 @@
 export * from "./sessionsView";
 export * from "./types";
-export * from "./mapObservationsTable";
-export * from "./mapTracesTable";
 export * from "./mapDashboards";
-export * from "./mapScoresTable";
 export * from "./promptsTable";
-export * from "./mapDatasetRunItemsTable";
